@@ -6,7 +6,7 @@ buildscript {
         mavenCentral()
     }
     dependencies {
-        classpath("org.springframework.boot:spring-boot-gradle-plugin:2.0.5.RELEASE")
+        classpath("org.springframework.boot:spring-boot-gradle-plugin:2.6.3")
         classpath('gradle.plugin.com.palantir.gradle.docker:gradle-docker:0.13.0')
     }
 }
@@ -16,6 +16,9 @@ plugins {
     id 'io.franzbecker.gradle-lombok' version '1.8'
     id 'java'
 }
+
+sourceCompatibility = 11
+targetCompatibility = 11
 
 apply plugin: 'java'
 apply plugin: 'eclipse'
@@ -42,15 +45,9 @@ docker {
     buildArgs(['DEPENDENCY': "dependency"])
 }
 
-
-
-
 repositories {
     mavenCentral()
 }
-
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
 
 dependencies {
     compile("org.hibernate:hibernate-entitymanager")
@@ -67,6 +64,5 @@ dependencies {
     compileOnly 'org.mapstruct:mapstruct-jdk8:1.3.0.Final'
     annotationProcessor 'org.mapstruct:mapstruct-processor:1.3.0.Final'
     testCompile("org.springframework.boot:spring-boot-starter-test")
-   
 }
 
