@@ -1,8 +1,11 @@
 package com.waracle.cakemgr.init;
 
-import com.waracle.cakemgr.exception.BusinessException;
 import com.waracle.cakemgr.entity.Cake;
+import com.waracle.cakemgr.exception.BusinessException;
 import com.waracle.cakemgr.repo.CakeRepo;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,10 +20,6 @@ import org.springframework.http.converter.HttpMessageConverter;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 /**
  * This class is to load initial data
