@@ -4,6 +4,6 @@ import com.waracle.cakemgr.entity.Cake;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CakeRepo extends JpaRepository<Cake, Long> {
-    Cake findByTitle(String title);
+
 }
 
