@@ -1,12 +1,11 @@
 package com.waracle.cakemgr.resource;
 
+import java.net.MalformedURLException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
-
-import java.net.MalformedURLException;
 
 @RestControllerAdvice
 public class CakeManagerAdvice {
