@@ -1,21 +1,23 @@
 package com.waracle.cakemgr.resource;
 
+import static org.springframework.util.Assert.notNull;
+
 import com.waracle.cakemgr.representation.CakeDto;
 import com.waracle.cakemgr.service.CakeManagerService;
 import io.swagger.annotations.Api;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.List;
-
-import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-import static org.springframework.util.Assert.notNull;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 
 /*
 This controller serves REST end points
@@ -31,7 +33,7 @@ public class CakeRestController {
     @Autowired
     private CakeManagerService cakeManagerService;
 
-    @RequestMapping(method = RequestMethod.GET, produces = {APPLICATION_JSON_VALUE})
+    @RequestMapping(method = RequestMethod.GET)
     @ResponseBody
     public List<CakeDto> getAllCakes() {
 
